+++ csv.go
@@ -2,10 +2,9 @@ package main
 
 import (
 	"fmt"
+	"os"
 	"path/filepath"
 	"time"
-
-	"github.com/ronbarrantes/woodchuck/utils"
 )
 
 // #### CSV FILE ####
@@ -42,7 +41,7 @@ func (f *CsvFile) InitCSV() error {
 	currDate := time.Now().Format("2006-01-02")
 	currentFileName := currDate + "-" + f.filename
 
-	err := utils.EnsureDirectoryAndFile(f.path, currentFileName)
+	err := EnsureDirectoryAndFile(f.path, currentFileName)
 	if err != nil {
 		return err
 	}
@@ -53,11 +52,51 @@ func (f *CsvFile) InitCSV() error {
 
 func (f *CsvFile) DeleteCSV() error {
 	fmt.Println("Deleting...")
-	err := utils.RemoveFile(f.fullpath)
+	err := RemoveFile(f.fullpath)
 	if err != nil {
 		return err
 	}
 
 	f.fullpath = ""
+	return nil
+}
+
+// EnsureDirectoryAndFile checks if a directory and file exist and creates them if they don't
+func EnsureDirectoryAndFile(path, filename string) error {
+	// Check if directory exists
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		fmt.Println("Directory does not exist, creating it...")
+		if err := os.MkdirAll(path, os.ModePerm); err != nil {
+			return fmt.Errorf("failed to create directory: %w", err)
+		}
+	}
+
+	// Combine directory and filename to get the full file path
+	filePath := filepath.Join(path, filename)
+
+	// Check if the file exists in the directory
+	if _, err := os.Stat(filePath); os.IsNotExist(err) {
+		fmt.Println("File does not exist, creating it...")
+		file, err := os.Create(filePath)
+		if err != nil {
+			return fmt.Errorf("failed to create file: %w", err)
+		}
+		defer file.Close()
+	}
+
+	return nil
+}
+
+// Remove File
+func RemoveFile(filepath string) error {
+	if _, err := os.Stat(filepath); os.IsNotExist(err) {
+		fmt.Println("Nothing to delete")
+		return fmt.Errorf("File %s doesn't exist", err)
+	}
+
+	if err := os.Remove(filepath); err != nil {
+		return err
+	}
+
 	return nil
 }
